Module faucon_emu::cpu

Falcon microprocessor abstractions.

Structs

Cpu
Representation of the Falcon processor.

CpuRegisters
Representation of all Falcon CPU registers.

Enums

CpuFlag
Flag bits for the flags special-purpose register.

ExecutionState
The execution state of the Falcon processor which controls its behavior.

Trap
Falcon trap kinds that can be delivered to the microprocessor.

Constants

CAUTH
A special-purpose register that holds the configuration bits for the SCP authentication process.

CX
A special-purpose register that holds the configuration bits for the SCP DMA functionality.

FLAGS
A special-purpose register that holds various CPU flag bits.

IV0
A special-purpose register that holds the address for Interrupt Vector 0.

IV1
A special-purpose register that holds the address for Interrupt Vector 1.

IV2
A special-purpose register that holds the address for Interrupt Vector 2.

PC
A special-purpose register that holds the current program counter.

SP
A special-purpose register that holds the current stack pointer.

TSTATUS
A special-purpose register that holds details on triggered traps.

TV
A special-purpose register that holds the address for the Trap Vector.

XCBASE
A special-purpose register that holds the external base address for IMEM transfers.

XDBASE
A special-purpose register that holds the external base address for DMEM transfers.

XTARGETS
A special-purpose register that holds the configuration bits for the CTXDMA ports.